Created Security class to be the base class of Stock and other security types. Refactored constructors to accomodate this.

+++ Asset Management Platform/StockList.cs
@@ -28,7 +28,7 @@ namespace Asset_Management_Platform.ViewModel
 
         public void AddStockToList(string name,
             string ticker,
-            decimal lastPrice,
+            double lastPrice,
             decimal bid,
             decimal ask,
             decimal beta,
